requirements.txt
paho-mqtt
cryptography
pyperclip
